Refactor pt. 1: move MPV player to own package and extract some types
